requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
email-validator==2.1.0
sqlalchemy==2.0.34
aiosqlite==0.20.0
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1